requirements.txt
openai>=1.12.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
python-dateutil>=2.8.2
python-dotenv>=1.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
pandas>=2.0.0
openpyxl>=3.1.0
fastapi[standard]>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
google-cloud-vision>=0.1.0
google-cloud-storage>=2.10.0
reportlab>=4.0.0
docx2pdf>=0.1.8
pypdf>=4.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.27.0

# LangGraph Agent Dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0